+++ editor/editor-common/src/__tests_external__/cases/base-test-collection.ts
@@ -1,3 +1,6 @@
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { context } from 'local-cypress';
+
 import { EditorTestCase } from './base-test-case';
 
 export interface EditorTestCollectionOpts {