+++ editor/editor-common/src/extensions.ts
@@ -13,6 +13,8 @@ export {
   getFieldSerializer,
   getFieldDeserializer,
   isFieldset,
+  isTabGroup,
+  isExpand,
   isDateRange,
   getUserFieldContextProvider,
   buildMenuItem,