+++ in-product-testing/in-product-testing-sample/__tests__/in-product/tests/editor.ts
@@ -1,3 +1,5 @@
+import { before, cy } from 'local-cypress';
+
 import { editorFundamentalsTestCollection } from '@atlaskit/editor-common/in-product';
 
 describe('Editor', () => {