+++ editor/editor-mobile-bridge/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @atlaskit/editor-mobile-bridge
 
+## 26.3.5
+
+### Patch Changes
+
+- Updated dependencies
+
+## 26.3.4
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 26.3.3
 
 ### Patch Changes