+++ in-product-testing/in-product-testing-sample/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @atlaskit/in-product-testing-sample
 
+## 0.1.1
+
+### Patch Changes
+
+- [`9e09b407b43`](https://bitbucket.org/atlassian/atlassian-frontend/commits/9e09b407b43) - Exclude `__tests_external__` from the `build/tsconfig.json`.
+  Add `local-cypress` and remove types export.
+
 ## 0.1.0
 
 ### Minor Changes