+++ media/media-card/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @atlaskit/media-card
 
+## 70.7.3
+
+### Patch Changes
+
+- [`9e09b407b43`](https://bitbucket.org/atlassian/atlassian-frontend/commits/9e09b407b43) - Exclude `__tests_external__` from the `build/tsconfig.json`.
+  Add `local-cypress` and remove types export.
+
 ## 70.7.2
 
 ### Patch Changes